hassio-to-pubsub-source/requirements.txt
google-cloud-pubsub==2.5.0
asyncws==0.1
